Titan-Analytics-Pro/requirements.txt
agno>=2.2.1
openai>=2.3.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
sqlalchemy>=2.0.0
duckdb>=1.0.0
fastapi>=0.110.0
uvicorn>=0.28.0
joblib>=1.3.0
